Add initial metrics collector
This is the start of the collector for adx-mon that is designed to
run on each node either on host or as a daemonset.  It will scrape
any pods that are scheduled on the same node as the collector and
that have opt-ed in to scraping via prometheus scrape annotation.

In addition to discovered targets, static targets can also be
specified to scrape well-known endpoints or host specific processes.
For example, etcd runs on masters, but not all agents.  A static
scrap target could be specfied to scrape the local metrics endpoint
only when on masters.

Scraped metrics are shipped to one or more remote write endpoints.
While multiple endpoints can be specified and the writes will
be replicated, there is currently retries or attempts to ensure
that failed writes are always replicated.  If one fails, those
samples are dropped from that endpoint.