Fix rule deletion requiring restart
When a rule has been removed, alerter should shutdown the worker
and stop executing the query.  This fixes a bug where we were not
shutting the worker down despite the rule being deleted.  It was
only stopped when alerter was restarted.  Since rules lifecycle
is managed outside of the alerter, we can't rely on restarting
alerter instances to stop rules.