Add basic notification throttling
This is a temporary measure to prevent notification storms due to
errant queries that return many rows for notifications.  This will
set the notification health metric to unhealthy, log an error and
cap the notifications as a configurable limit.

Next iteration of this will send a notification to the rule destination
indicating that the alert is throttled.