Switch notifcation health metric to unehalthy metric
Notification health metric was not working well because the metric
is transient and resets when the process restarts.  If an alert is
setup to watch for a drop to zero, it would fire when the process
restarts.

Instead, we'll switch this to be a metric indicating unhealthiness
when true and healthy and false.  This allows a process restart
to reset it back to healthy, but if problem persists, it will
get reset back to true to indicate unhealthiness.