Send alert to rule destination when query fails
If there is an issue with the query, an unhealthy metric is signal
which indicates a problem.  The owner of the alert likely won't see
this metric and it takes time to dig into what the root cause was.

If it's a query error, we'll send an ICM with the error message to
the query destination so the owner can see this more quickly and
address the issue.  If we fail to create that ICM, we'll set the
unhealthy metric so that the adx-mon operator can be alerted to
the problem.