Fix double targets getting added
The informer was started before we setup the initial list of targets.
Since the informer runs in a separate goroutine, it raced with the main
thread setting up the initial set of targets.  This could cause some
targets to get double added by mistake.

This move the informer startup to after the intial target discovery
as well as simplifies the target slice locking to reduce the chance
of a race in other contexts.