Add explicit listen addr for collector in tests
The start process for the collector service includes a goroutine that
binds to a configured port to expose prometheus metrics. The default
value is an empty string, causing the Go http server to try binding to
:80. This makes some of the service tests flaky as it sometimes manages
to run the server bind goroutine before the test finishes, leading to a
permission denied error.