Fix scrape target discovery to scrape container ports
The initial discovery was not right in that it was only creating
one target per pod which is not correct if there are multiple
containers within the pod.

This changes the discovery to create a target per container when
opt-ed in via scrape labels.  If a scrape port is specified, we'll
just scrape the one port on the pod instead of all the container ports.