Add token authentiation method
Alerter in the future may need to support a number of Kusto
authentication methods in the future to support interactive use by
developers, CI usage, and others. This adds a more generic mechanism to
provide a list of possible auth mechanisms which fall back to latter
methods if the current method is unusable. These decisions are currently
based on what arguments are provided and do not exercise the
authentication method to make the decision.