Require scoping metadata for static scrape targets
These scrape metrics are not that useful at query time without
some context of where they originated.  This requires a value to
be specified for namespace, pod and container that needs to be
provided by the user.  Blank values are allowed if the user understands
that querying these metrics might not make much sense.