Add alert criteria/tag support
This adds a new feature to provide better control of where an
alert can execute.  In some envs, there may be multiple envs, clouds,
regions, or topologies that make sense to the user.  Alert may not
always be able to execute in all envs.  For example, some alerts
may call to Kusto database that are not reachable from other clouds.

This feature adds the a new `criteria` map to the AlertRuleSpec
definition that allows for a map of key/value pairs.  Each pair
defines a criteria that must be true for the rule to execute in
a given env.

Alerter instances specify tag key/value pairs as command-line options
that alerts can be configured to use.  Since region and cloud
flags are already required, these flags are added as tags by
default and new tags to define a users environment can be added
in addition.