Fix lost peer change update
This is an old informer issues where a delete and add or add and delete
could get converted to an update call which messes up the peer syncing.
This causes ingestors to sometimes get stuck with old addresses and
accumulate old segments that it thinks it should transfer.  This
fixes the first part of this so that the address eventually are kept
in sync.