Add drop metrics/labels support to collector
This adds the same capability as ingestor to collector to drop
labels and metrics matching regexes.  This is useful to drop data
that is not desirable to store from even being collected.  Doing
this as collection time can reduce resource usage on collectors and
reduce the volume of data that ingestors and Kusto needs to process.