OTLP API requires a protobuf response
When sending a response from LogsProxyHandler, we must do so as a
Protobuf, even upon success. In the case of a partial success, we
need to send then appropriate ParitalResponse Protobuf but still send
a StatusOK. Since we're proxying to more than one upstream receiers, we need
to determine the maximum number of rejected logs to be represented within
our partial response so that the caller can make the appropriate decision.