Close idle connections periodically to balance load through SLB
Collector connections (and other clients) use HTTP keep-alive which
prevents connections from being closed on the server.  This can cause
problems when scaling out ingestor pods because the existing connections
persist and are not balanced to new pods.

This decreases the idle timeout to forces connections to close so a new connection
can be established which can land on a different host.