Adjust batches to handle large segment sizes better
In some cases the segments are very large and accumulate quickly.
The batch was not breaking these cases down into multiple batches
which can cause the uploader to try and upload one huge file and
get backlogged.

Instead, we break the batches up into multiple batches to allow for uploading
in parallel.