Switch prioritization of segment uploads
This switches the segment uploads to prioritize newer data over
older.  Since monitoring data generally alerts on recent data,
if we are lagging significantly, it can take a long time to catch
up if we upload oldest first.

Instead, try to keep most recent data uploaded and if backfill
old data that we missed in the background.