Adjust upload priority ordering
The prior change to upload only the newest only works if we can
fully drain the backlog.  If we can't get past the backlog, the
oldest entries stay on disk indefinitely.

Instead, still prefer newer data, but add the last 10% of old
segments to the front of the queue so we can always chip away
at the backlog a little.