Replace cardinality counts with kusto queries
The cardinality counting approach doesn't work with multiple counters
as the counts are not mergeable.  We would need to insert the kusto
HLL format and merge that at query time.  Since that format is not
docuemented publicly nor is the specific implementation available
to be used like this, we have to rework how counts are tallied.

Instead, just run a function that dcounts all series IDs on all
tables and store that as a new metric.  Only one ingestor will
run this query periodically.