Ensure max-batch-size is enforced
The max-batch-size controls how big of a batch to send in each
write request.  This was not completely envforced so it was possible
to send batchs much larger than max-batch-size depending on the
kinds of metrics being scraped.

This changes the collector to ensure a batch never exceeds the
max-batch-size now.