Add backpressue signaling
When a node because backlogged with too many uploads and transfers
it can struggle to catch up if it continues to access new writes
and new transfers from peers.  In these cases, it's preferable to
reject writes and transfers to allow the node to catch up.

This adds a back pressure signaling mechanism to tell peers to
stop transfering segments to this peer and upload them directly
for the next minute.  This will also reject incoming writes from
clients, but they can't directly control not sending new writes
to this instance.

A subsequent PR will tie into this health check mechanism to fail
a readiness probe so that the node can be removed from the load
balancer temporarily.