Move logging to slog
This involved swapping out the logging implementation and removing the
limited usages of logger instances. We just use the default global one
now.

This change also allowed go vet to catch a couple of incorrect usages of
formatter methods including args that were in the wrong order compared
to the format string.