Delete old segments that do not have a known destination
In cases such as OLTP logs, the sender defines the destination database.
Due to misconfiguration or badly-behaved clients, sometimes ingestor may
receive traffic destined for ADX clusters that have not been configured
in ingestor. We will retain these segments and keep attempting to upload
them in cases where we may receive data for a destination prior to
ingestor being configured for it.

In the face of continuous inability to find a destination for these
logs, we need to expire them so they do not continue to grow or throw
off metrics for max segment ages.