Reject unknown log destinations, accept logs that we can write
This makes a couple of changes to the log ingestion process. First, we
now ensure that ingestor actually knows about a log destination database
prior to accepting these logs. Second, in cases where some logs are
missing routing information or have unconfigured destinations in their
attributes we continue to consume the logs that we can, then inform the
sender at the end that were were some invalid logs sent. This prevents
some new or unknown logs from impacting the delivery of other known and
valid logs.