Creating initial log collection framework
This creates some scaffolding for log collection in collector. This
implementation is intended to collect from one or several sources and
feed them to ingestor via otel.

The collector/logs/types package is intended as a library that can be
vendored into later Go-based plugins. As a result, it is intended to be
relatively independent and require a minimum of other dependencies.