Change table partitioning strategy
The partition of tables across nodes currently assigns one node
to be the owner of any given table to be uploaded.  This allows
all the small writes to accumulate onto distinct nodes and balance
the tables across all nodes.

Unfortunately, this strategy breaks down when the number of nodes
is greater than the tables in that some nodes will not be assigned
any tables.  This strategy can have issues with really large tables
as the load will always be limited by one node.  Finally, as the
total number of nodes grows, there is a very large fan-in that happens
to all nodes where all the peer are transferring their files to the
owner.

To address this, instead of partition all the keys across all nodes,
 we split the total nodes into sets of nodes where the keys are partitioned
across each set.  This can lead to slightly less optimal uploaded file
sizes, but allows all nodes in the cluster to better handle teh situations
described above.

In the first case, each group of nodes will have an owner for a given
key.  As the pool size grows, more groups are created.  This allows
all nodes to have some ownership of the key space.

In the second case, large tables are split across different groups of
nodes so multiple nodes can take part of the ownership.

In the last case, the fan-in in bounded by the size of the group.  If
the group is sized to a max of 10 nodes, each node will have a max fan-in
of 10 regardless of how large the cluster grows.