Handle metrics with special characters
Some metrics may have : or other characters that are valid
file names, but not legal Kusto entity name characters.  When these
are sent, ingestor gets stuck with the segments on disk and can never
complete the upload.

This fixes the Normalize function to strip out all invalid characters
so that only [a-zA-Z0-9] are present.  For importing segments, the
segemnt file will be rejected so that ingestor doesn't get into this
state.  On the collection side, the metric names will be normalized
to allowed naming conventions before storing in the WAL.