Add experimental log reading support for journald
This is an initial implementation to read logs from journald to help
refine the log collection APIs and interfaces. Currently this reads
all streams in the journal and persists a file on disk to save the
current read position, then starts from that position if possible.

This is only enabled with the --experimental-log-collection flag.