Track segments already part of a batch via batcher
The batcher would just group all segments on disk into batches
even if the segments were already part of a batch in flight. This
caused some spurious errors in the logs about segments not existing
because they were processed as part of another batch.

This changes the batcher to keep track of segemnts that are part
of other batches so we can avoid this errors.