Add batch transferring of segments
Segment files can be appeneded together to reduce the number
of files provided that the internal schema, DB, table are all
the same.  This change updates the replicator to transfer
all the segments in a batch together so they are combined
on the destination host during transfer.  This is similar to
what happens on the upload to kusto path where multiple segments
are combined into one larger file for Kusto to ingest.

This has a few benefits:

1) Reduces back-pressure on the destination node as they will
end up with fewer files on disk and need to upload as many.
2) Reduced disk/network IOPS since multiple files are combined
into a single larger file.
3) Reduce error logging when one transfer fails as the either
the whole batch succeeds or fails now.