Switch Import to append to latest segment
When there are many small files, the IOPS due to
creating, fysnc and renaming files is too high since
it's multiple by the number of files being imported.

Switching import to append to the active segements allows
this code path to leverage the buffering and optimizations
in place for heavy writes.