Move index initialization to Repository
This moves the initialization of the index to the repository so
it can be done once instead of N times (each WAL).  When a WAL
is opened, it will just create a new segment and allow the existing
segments to be added to the index.

This should help with startup times and large segment directories.