Support drop all metrics by default in collector
Currently, collector accepts all metrics and requires drop rules
to drop unwanted metrics.  It is sometimes desirable to invert this
and drop everything except metrics that are allowed.

This add supports for the latter case to allow metrics by regex as well
as subsets that may be interesting based on certain label key and
value paires.  For example, you might want to collect container
disk metrics but only for a specific container or valume and not all of them.