Switch bytes pool to a more memory efficient implementation
Our current bytes pool has a few issues.  The first is that when
the size of the buffers varies substantially, the buffers gradually
all trend towards the largest size needed.  For example, if there
are many small 8k buffers needed, but occassionally one 1MB one needed,
they will all eventually grow towards 1MB and we waste a lot of memory.

The second issues is that we have many independent pools allocated and used
by different parts of the code.  This also wasts memory since not all of
them are used all the time.

This switches our pool to a better implementation that handles tiering
of pools with a shared global pool.  The first issues is addressed by
the tiering that allows lots of 8k buffers to be held while only one 1MB
buffer is held.  This avoids the wasted memory.  The second issue is address
by wrapping this library with our impl so that any pools created in the current
code actuall re-use the same backing pool.