Update segment merging to account for wal headers
The prior segment merging implementation just concatenated
all the segments together which still creates a valid segment
file.  With the addition of the segment headers, the merge
files are no longer valid because the segment header is added
where the block headers should be.

This reworks the merger to account for the segment headers by
first prepending the segment header and then stripping the
segment header from all merged files.