Retry writes one more time if segment closes
The WAL can rotate a segment while a write is inflight but not
actually sent to the segment yet.  In this cases, the write fails
with ErrSegementClosed which goes back to the caller to retry.

This causes some noise in the logs and extra requests when we could
make this transparent to the caller by retrying the write on the
new segment.