Fix batching with filtering metrics
When default drop all and allow/drop rules are in place, the batches
can be smaller than the mast batch size and we end up sending multiple
smaller request to ingestor.  This fixes that case and fills up each
batch to the max batch size as before resulting in few requests that
are generally the same size.