Filter metric names directly rather than using label/metric value
Evaluating two regexes per loop looking for the name of the metric and
the value ends up causing a lot of CPU usage for the cases when we
filter most metrics out from a scrape. In particular, the old method of
using '^__name__\\b' could not be evaluated in one pass, so a lot of
time was spent in the regex backtracking process.

Instead, keep the allowed metric names separated and check for the
matches directly with the name we've already looked up to save a lot of
looping and regex-reruns.