Set top-level config for pod discovery for tailog
Within pods, we do not consume a kubeconfig file but instead use kube
libraries to get a token within the pod filesystem. Instead of using the
kubeconfig configuration as a proxy for doing pod discovery, assume we
wish to unless it is explicitly turned off with disable-kube-discovery.

Only one tail-log instance can have disable-kube-discovery == false, as
it doesn't make sense to have multiple tail-log configs scraping the
same set of pods.