Set max query time, track number of total queries
We want to ensure we never get stuck running queries so we continue to
make forward progress, even in the face of stuck network connections and
the like.

We now emit a counter to track the number of queries being run over time
to more easily monitor for the condition where queries are not
successfully running.