Emit latency metrics (#340)
When a log is received in Collector, we emit metric collector_logs_latency, which is the amount of time elapsed, in seconds, between when the log was generated and when it was received by Collector. These logs are turned into WAL segments in Collector, batched, then sent to Ingestor. In the process of creating WAL segments, the filename contains a flakeid with an epoch of the time the file was generated. In Ingestor, prior to uploading the file to Kusto, we inspect the filename's flakeid, convert it to the epoch time of when the file was created, then compare this with the current time, this is then emitted as the metric ingestor_sample_latency, also in seconds. We can then use these to metrics, in addition to the Kusto generated metric "ingestion_time", to determine the total end-to-end latency and the latencies at each touch-point in time.