Change segment iterator to read as much as possible
If there is a segment with 0 values at the end, it can cause
imports to fail because the checksum fails for the block. This
can cause collector to get stuck on this file.  A repair of the
segment would truncate this block which would cause the import
side to not see the block.

In this case, we can treat the segment as if it was repaired and
just stop iterating on the segment when the block len is 0.  This
will allow all the valid data to be imported and bad data to be
dropped.  In the case where the trailing block is actually corrupted
and does not have a 0 length, it will retrn an error.