Fix stale scrape targets issue
There is a chance that pods can change IPs if the CRI restarts or
kills a pod behind the scenes.  This can cause another pod to be
assigned an IP that collector is scraping and either scrape the
wrong things or continue to fail.

This improves the detection of when a scrape target changes.