Add option to disable target discovery
This allows collector to be used for just static scraping of specific
targets.  Since collector is intended to be deployed as a DaemonSet,
it's not idel to scrape things like the api server, service endpoints,
etc.. from all the nodes since that can create lots of duplicae data.

This option allows target discovery to be disabled so that you can
deploy a separate collector with fewer replicas that scrapes static
targets only.