Rework collector pod discovery
The pod discovery with the informer wasn't working reliably.  Sometimes
pod delete events would get missed so a collector instance would continue
scraping the endpoing and get "no route to host" errors.  The informer
should be resyncing every minute, but it's not working.

To address this, we have each collector list pods every minute to resync
and reconcile targets.  Since collector is a daemonset and list pods from
every node could be expensive, we take advantage of the fact that we only
need the list of pods on the current nodes.  Conveniently, kubelet exposes
a pods API that returns the pods on the current nodes.

This change scrapes the local pods API to reconcile pods on the node which
avoids taxing the API server/Etcd and serializing the pods over the network.