Log reason for ingestor instance reporting unhealthy
There are a few limits where ingestor will start to throttle client
requests.  For example, if the number of segments exceeds the max
segment limit or the size of segments on disk excees the max disk
usage limit, ingestor will reject new writes to trigger back pressure
to clients.

When these occur, it's hard to known the reason why an instance is
reporting unhealthy unless you are familiar with the various health
checks ingestor uses internally.

This adds the reason to the health logging so we can make it easier
to determine what remediation might be necessary to get instance healthy
again.  For example, if max segments is exceeded, a DRI might need to
increase the max segments limit if the reported value is expected.