Handle bad request replication errors by dropping the segment
If collector or ingestor gets a bad request response from ingestor,
this is a fatal error for the client as it will continue to resend
the same request over and over.

To prevent this from repeating indefinitely, we log and drop the
segment as it's not able to be transferred.