Switch metric fetching to use streaming approach
When scraping targets, some targets can cause memory usage in
collector to spike.  This causes the majority of pods to have low
memory usage and a few that are several orders of magnitude higher.
To account for this, we have to set higher limits to prevent these
pods from getting OOM killed.

This occurs becauset he metrics scraping pulls down all the metrics
into a buffer, then that is parse and converted to a big map of
prometheus types and then that is subsequently converted to lists
of protobufs to be shipped to ingestor.  For big requests, the memory
spikes.

To address this spike, we use a streaming approach where the metrics
payload is stream and each series is read one at a time and converted
directly to the protobuf that will be sent to ingestor.  When we have
enough series in a batch, it is flushed.  This avoids the big memory
spike and uses a relatively constant amount of memory regardless of
metrics returned from scrape targets.

A side effect of this is that ingestor CPU utilization and disk IO somehow
goes down as well.  I can't fully explain this yet, but I suspect it
has something to do with the removal of the intermediate map which
causes series to lose their ordering.  I suspect this may cause
poor compression and more lock contention on ingestor as all the
batches are unordered and writing/acquiring locks more randomly.