Reduce lock contention in WAL
In the write path, a Write lock is needed in the slow path to ensure
a segment exists.  This was getting hit more frquently than expected
because we roll segments more quickly and the max segment size triggers
a rollover more frequently.

To address, this we create a new segment when we rotate proactive so
that the slow patch for writes isn't hit as oftent.  The side effect
of this is that if writes not frequent, we roll over emtpy segments
frequently.  If we try to upload them, they show up as ingestion failures
in kusto because the file has no rows.  To address this issue, we just
remove the segment if it only contains the segment magic header bytes.