Create table/schema mapping from embedded segment schemas
If the segment cotains a schema, we use that to upload the segment
versus using the default schema.  This change is backwards compatible
so segments that do not contain a schema still use the default schema.

The inten with this is so we can move the lifting of columns and things
to collector and make ingestor fairly simple.  Right now both ingestor
and collector have duplicate functionality which isn't lining up for
various uses.