Change kubernetes log tailing to use /var/log/containers
This allows us to handle several situations more cleanly. In situations
where containers within a pod crash but do not replace the pod itself,
new containers are created with new container ids. This keeps the same
top-level directory for the pod within /var/log/pods, but it may create
entirely new paths within the directory that do not necessarily follow a
0.log file name convention. Instead, sometimes new files are created as
1.log, etc.

In these cases, the source of truth for the correct file is the file
associated with the correct container id within /var/log/containers
ending in .log. This appears to be stable even in cases where containers
are recreated inline.

This also fixes issues where we were not looking at the correct
ContainerStatuses field for init and ephemeral containers for the
container id.