Don't release WriteRequest to the pool in proxy
The metrics proxy was incorrectly adding the WriteRequest back to
the pool since it was done wit it.  This is usually the right path
but in the proxy path, the writes are buffered and flushed in batches
so the request is not actually void of uses.  This causes a panic
and label corruption if the request and timeseries are reused again
while they are cached.