Fix unmarhalling labels
The easyproto.FieldContext.Bytes() notes that that returns a []byte
from the underlying buffer and it's only valid while the buffer
has not changed.  This buffer is coming from a pool and we release
it before this use was completed.  This causes the labels to get
overwritten by the next user of the pool under higher concurrency.