Better handle missing containerids
There are cases when we are notified of a new container and there is a
ContainerStatus object that matches the container, but the container is
not running yet so it does not have a containerID property within it
yet. This messes up our state machine where we attempt to tail the wrong
file without a containerID as part of the file path.