Rework http connection closing
Due to a golang bug related to http2 connections not getting
closed when they are broken at times, we disabled http2 and also
disabled long-lived connections.  While this resolved the issues
we saw, ingestor and collector have much higher CPU due to the
constant TLS connection setup and teardown.

Since we understand the root problem better now, this re-enables
persistent connections with http 1.1 with a gradual drain on
the server side.  The drain is used to when scaling out so that
traffic eventually becomes balanced.  This also add a mitigation
for http2 (though still disabled) which would allow us to use http2
connections in the future.

http2 is still disabled because even when we send a GOAWAY message
from the server, we still get spurious client side errors.  When
the upstream bug is fixed, we can revisit http2 enablement.