Make segment flush interval configurable
The segment flush interval defaults to 100ms which allows time
for many small writes to fill up the buffer and be flushed to disk
in larger batches.  When there a lot of metrics or tables and not
a lot of concurrent writes, this flush interval increases CPU utilization.

This setting allows for increasing the flush interval in collector to
reduce CPU usage since the number of writes is relatively infrequent.
For ingestor, we leave the default value in place.