KQL Function CRD (#419)
* Introduce a CRD for managing KQL functions and views

* Create KQL Function store

* Create a sample View

* Periodically sync functions

* Update quickstart bundle

* Rework http connection closing

Due to a golang bug related to http2 connections not getting
closed when they are broken at times, we disabled http2 and also
disabled long-lived connections.  While this resolved the issues
we saw, ingestor and collector have much higher CPU due to the
constant TLS connection setup and teardown.

Since we understand the root problem better now, this re-enables
persistent connections with http 1.1 with a gradual drain on
the server side.  The drain is used to when scaling out so that
traffic eventually becomes balanced.  This also add a mitigation
for http2 (though still disabled) which would allow us to use http2
connections in the future.

http2 is still disabled because even when we send a GOAWAY message
from the server, we still get spurious client side errors.  When
the upstream bug is fixed, we can revisit http2 enablement.

* Add missing repair step for corrupted segment

This was a case that was missing a repair when we get an error
while reading the block.

* Make segment flush interval configurable

The segment flush interval defaults to 100ms which allows time
for many small writes to fill up the buffer and be flushed to disk
in larger batches.  When there a lot of metrics or tables and not
a lot of concurrent writes, this flush interval increases CPU utilization.

This setting allows for increasing the flush interval in collector to
reduce CPU usage since the number of writes is relatively infrequent.
For ingestor, we leave the default value in place.

* Fix comment

* Add support for multi-region AKS-ADX setup

Until now bundle.sh would expect that the AKS and ADX clusters be in the
same region. This PR updates the script to add support for these
clusters to be in different regions.

* Set function status object

* Update CRD to allow arbitrary KQL execution

* Remove invalid condition from testcase

* Update function definition to reflect raw KQL statements

---------

Co-authored-by: Jason Wilder <jawilder@microsoft.com>
Co-authored-by: Tanay Kothari <tanaykothari@microsoft.com>