Don't block appends when segments are highly contented
Under higher write load, the segment write lock can cause
long delays for a goroutine to acquire it for a write and append
operation.  This causes ingestor to report "slow request" at times
because multiple locks are getting blocked for while and the request
goes slower.

For collector shipping segments, it generally has many segements
on disk that it needs to send.  So instead of blocking when a segment
is locked, we now just return a locked error and let collector move
to the next segment.  It will retry the prior segment again in a few
seconds.  This allows for better and reduces the slow requests warnings
quite a bit.