Ingestor support for Kustainer
Kustainer has several limitations that impact Ingestor's interaction with Kusto.
- Kustainer has no auth
- Kustainer only supports inline ingestion

To support the lack of auth, this change inspects the Kusto cluster URL and if it's an insecure connection, we do not attempt to add auth to the Kusto connection object.

To support inline ingestion, we query Kusto's cluster details and if the cluster is marked as "KustoPersonal", we use this as a trigger to switch into an inline ingestion mode.