Update function tasks to extract specific errors and cache successes
We now update our function cache when we have successfully created a
function. This allows us to skip repeated reconciliations when the
function CRD has not changed. We also evaluate if the Generation has
changed rather than the ResourceVersion. Generation is updated when our
definition changes, while ResourceVersion captures all changes including
status updates which would lead us to continually applying and updating
the same CRDs.

When updating the CRD's status object, we now attempt to extract a Kusto
HttpError from the error object. If it exists, we extract the specific
"@message" field to get a more specific error message to place into the
status field. Otherwise, we default to the return of Error().