Update containerd to 1.7.24
Various scanners are upset about
go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc
v0.45.0. This is not depended on by us and only shows up in go mod
graph as a dependency of containerd, but not go list. Upgrading to the
latest patch version of the 1.7 series pulls in the fixed version.