Implement KQL CRD informer to better handle deletions
This PR introduces dead code that will replace the existing listing mechanism for KQL CRDs. We want to be able to react to CRD deletions by deleting the associated resource in Kusto; however, by simply listing the CRDs on a set cadance leaves us with a lot of ambiguity about how we properly identify resources that have been deleted. By using an informer, we can safely distinguish between create, update and deletion events and respond accordingly. This PR also leaves open the question of leader election. It's not safe to have many actors mutating Kusto resource state. We can either use leader election or the partitioning schema used for batch ownership to limit the number of actors mutating the same resources in Kusto.