Add "suspended" and "observedGeneration" to the funtions CRD
The suspended property enables an operator to suspend processing of a CRD during some investigative period.

The ObservedGeneration property enables us to remain stateless with respect to our processing model. So instead of having to maintain a cache so that we can determine when a CRD has been updated, we can remove the cache and know if a CRD requires reconciling by comparing its ObservedGeneration to the current Generation.