Combine the functionality of storage and crd
There is an extra bit of unecessary abstraction that complicates our Function handling where our storage package is acting as a glue between the crd and adx packages. By combining the storage and crd package it's far easier to understand the code's flow. Furthermore, the code that lives in pkg is generally meant to be shared by more than one other package, so until that's necessary, having a crd package doesn't fit the intent of pkg.