Integrate new function store to task management
Several unit tests were removed for this commit due to the shift of responsibility with respect to ObservedGeneration and LastReconcileTime, which was moved from task management to the function store layer. The reason for moving the responsibility for these two fields is that it's a bit of an implementation leak for the kusto layer to have to understand how and why the ObservedGeneration and LastReconciledTime are utilized.